put code in folders, create own file for FmuServiceHandler
