+++ cpp/FMU-proxy/CMakeLists.txt
@@ -3,16 +3,16 @@ project(FMU-proxy)
 
 set(CMAKE_CXX_STANDARD 17)
 
-include_directories(thrift-gen)
+include_directories(common/thrift-gen)
 
 add_executable(FmuTest
-        thrift-gen/definitions_types.cpp
-        thrift-gen/definitions_types.h
-        thrift-gen/service_types.h
-        ThriftHelper.h
-        FmuWrapper.cpp
-        FmuWrapper.h
-        FmuTest.cpp)
+        common/thrift-gen/definitions_types.cpp
+        common/thrift-gen/definitions_types.h
+        common/thrift-gen/service_types.h
+        common/ThriftHelper.h
+        server/FmuWrapper.cpp
+        server/FmuWrapper.h
+        test/FmuTest.cpp)
 
 target_link_libraries(FmuTest fmilib)
 target_link_libraries(FmuTest boost_system)
@@ -20,15 +20,15 @@ target_link_libraries(FmuTest boost_filesystem)
 target_link_libraries(FmuTest dl)
 
 add_executable(ThriftServer
-        thrift-gen/definitions_types.cpp
-        thrift-gen/definitions_types.h
-        thrift-gen/service_types.h
-        thrift-gen/FmuService.cpp
-        thrift-gen/FmuService.h
-        ThriftHelper.h
-        FmuWrapper.cpp
-        FmuWrapper.h
-        ThriftServer.cpp)
+        common/thrift-gen/definitions_types.cpp
+        common/thrift-gen/definitions_types.h
+        common/thrift-gen/service_types.h
+        common/thrift-gen/FmuService.cpp
+        common/thrift-gen/FmuService.h
+        common/ThriftHelper.h
+        server/FmuWrapper.cpp
+        server/FmuWrapper.h
+        server/ThriftServer.cpp server/FmuServiceHandler.h)
 
 target_link_libraries(ThriftServer thrift)
 target_link_libraries(ThriftServer fmilib)
@@ -39,12 +39,12 @@ target_link_libraries(ThriftServer dl)
 
 
 add_executable(ThriftClient
-        thrift-gen/definitions_types.cpp
-        thrift-gen/definitions_types.h
-        thrift-gen/service_types.h
-        thrift-gen/FmuService.cpp
-        thrift-gen/FmuService.h
-        ThriftClient.cpp)
+        common/thrift-gen/definitions_types.cpp
+        common/thrift-gen/definitions_types.h
+        common/thrift-gen/service_types.h
+        common/thrift-gen/FmuService.cpp
+        common/thrift-gen/FmuService.h
+        client/ThriftClient.cpp)
 target_link_libraries(ThriftClient thrift)
