+++ cpp/FMU-proxy/client/ThriftClient.cpp
@@ -30,8 +30,8 @@
 #include <thrift/transport/TTransportUtils.h>
 #include <thrift/stdcxx.h>
 
-#include "thrift-gen/FmuService.h"
-#include "thrift-gen/definitions_types.h"
+#include "../common/thrift-gen/FmuService.h"
+#include "../common/thrift-gen/definitions_types.h"
 
 
 using namespace std;
