+++ cpp/FMU-proxy/common/ThriftHelper.h
